+++ src/backend/interfaces/REST/go.server/errors.go
@@ -1,5 +1,12 @@
 package rest_api
 
+import (
+	"github.com/CaliOpen/Caliopen/src/backend/interfaces/REST/go.server/middlewares"
+	swgErr "github.com/go-openapi/errors"
+	"gopkg.in/gin-gonic/gin.v1"
+	"net/http"
+)
+
 type (
 	Errors []Error
 
@@ -24,3 +31,9 @@ type (
 		Message string `json:"message,omitempty"`
 	}
 )
+
+func notImplemented(ctx *gin.Context) {
+	e := swgErr.New(http.StatusNotImplemented, "not implemented")
+	http_middleware.ServeError(ctx.Writer, ctx.Request, e)
+	ctx.Abort()
+}