+++ src/backend/interfaces/REST/go.server/middlewares/authentication.go
@@ -6,23 +6,15 @@ package http_middleware
 
 import (
 	"encoding/base64"
-	"encoding/json"
+	"github.com/CaliOpen/Caliopen/src/backend/main/go.backends"
 	"gopkg.in/gin-gonic/gin.v1"
-	"gopkg.in/redis.v5"
 	"net/http"
 	"strconv"
 	"strings"
 	"time"
 )
 
-type auth_cache struct {
-	Access_token  string    `json:"access_token"`
-	Expires_in    int       `json:"expires_in"`
-	Expires_at    time.Time `json:"expires_at"`
-	Refresh_token string    `json:"refresh_token"`
-}
-
-func BasicAuthFromCache(cache *redis.Client, realm string) gin.HandlerFunc {
+func BasicAuthFromCache(cache backends.APICache, realm string) gin.HandlerFunc {
 	if realm == "" {
 		realm = "Authorization Required"
 	}
@@ -37,20 +29,15 @@ func BasicAuthFromCache(cache *redis.Client, realm string) gin.HandlerFunc {
 		if user_id, access_token, ok = BearerAuth(c.Request); !ok {
 			if user_id, access_token, ok = c.Request.BasicAuth(); !ok {
 				kickUnauthorizedRequest(c, realm)
+				return
 			}
 		}
 
-		// Search user in redis cache of allowed credentials
-		var cache_str string
-		var err error
-		cache_str, err = cache.Get("tokens::" + user_id).Result()
-		if err != nil {
+		// Search user in cache of allowed credentials
+		auth, err := cache.GetAuthToken("tokens::" + user_id)
+		if err != nil || auth == nil || auth.Access_token != access_token || time.Since(auth.Expires_at) > 0 {
 			kickUnauthorizedRequest(c, realm)
-		}
-		var cache auth_cache
-		err = json.Unmarshal([]byte(cache_str), &cache)
-		if err != nil || cache.Access_token != access_token || time.Since(cache.Expires_at) > 0 {
-			kickUnauthorizedRequest(c, realm)
+			return
 		}
 
 		//save user_id in context for future retreival
@@ -62,28 +49,6 @@ func kickUnauthorizedRequest(c *gin.Context, realm string) {
 	// we return 401 and abort handlers chain.
 	c.Header("WWW-Authenticate", realm)
 	c.AbortWithStatus(401)
-}
-
-//bespoke unmarshaller to workaround the expires_at field that is not RFC in cache (tz is missing, thanks python)
-func (ac *auth_cache) UnmarshalJSON(b []byte) error {
-	var temp struct {
-		Access_token  string `json:"access_token"`
-		Expires_in    int    `json:"expires_in"`
-		Expires_at    string `json:"expires_at"`
-		Refresh_token string `json:"refresh_token"`
-	}
-	if err := json.Unmarshal(b, &temp); err != nil {
-		return err
-	}
-	ac.Access_token = temp.Access_token
-	ac.Expires_in = temp.Expires_in
-	expire, err := time.Parse(time.RFC3339Nano, temp.Expires_at+"Z")
-	if err != nil {
-		return err
-	}
-	ac.Expires_at = expire
-	ac.Refresh_token = temp.Refresh_token
-	return nil
 }
 
 func BearerAuth(r *http.Request) (username, password string, ok bool) {