+++ src/backend/main/go.backends/NotificationsInterfaces.go
@@ -8,10 +8,10 @@ import (
 	. "github.com/CaliOpen/Caliopen/src/backend/defs/go-objects"
 )
 
-//LDA only deals with email
 type NotificationsStore interface {
 	CreateMessage(msg *Message) error
 	UserByUsername(username string) (user *User, err error) // to retrieve admin user
+	GetLocalsIdentities(user_id string) (identities []LocalIdentity, err error)
 }
 
 type NotificationsIndex interface {