+++ src/backend/main/go.backends/store/cassandra/users.go
@@ -9,8 +9,8 @@ package store
 import (
 	"errors"
 	. "github.com/CaliOpen/Caliopen/src/backend/defs/go-objects"
-	"github.com/gocql/gocql"
 	"github.com/gocassa/gocassa"
+	"github.com/gocql/gocql"
 )
 
 func (cb *CassandraBackend) RetrieveUser(user_id string) (user *User, err error) {
@@ -36,7 +36,7 @@ func (cb *CassandraBackend) UpdateUser(user *User, fields map[string]interface{}
 	return userT.Where(gocassa.Eq("user_id", user.UserId.String())).Update(fields).Run()
 }
 
-func (cb *CassandraBackend) UpdateUserPassword(user *User) error {
+func (cb *CassandraBackend) UpdateUserPasswordHash(user *User) error {
 	return cb.Session.Query(`UPDATE user SET password = ? WHERE user_id = ?`,
 		user.Password,
 		user.UserId,
@@ -67,4 +67,15 @@ func (cb *CassandraBackend) GetLocalsIdentities(user_id string) (identities []Lo
 	}
 
 	return
+}
+
+// UserByRecoveryEmail lookups table user_recovery_email to get the user_id for the given email
+// if a user_id is found, the user is fetched from user table.
+func (cb *CassandraBackend) UserByRecoveryEmail(email string) (user *User, err error) {
+	user_id := new(gocql.UUID)
+	err = cb.Session.Query(`SELECT user_id FROM user_recovery_email WHERE recovery_email = ?`, email).Scan(user_id)
+	if err != nil || len(user_id.Bytes()) == 0 {
+		return nil, err
+	}
+	return cb.RetrieveUser(user_id.String())
 }