+++ src/backend/main/go.main/users/password.go
@@ -28,6 +28,11 @@ func ChangeUserPassword(user *User, patch *gjson.Result, store backends.UserStor
 	}
 
 	new_password := patch.Get("password").Str
+
+	return ResetUserPassword(user, new_password, store)
+}
+
+func ResetUserPassword(user *User, new_password string, store backends.UserStorage) error {
 	//compute new password strength
 	user_infos := []string{user.Name, user.GivenName, user.FamilyName, user.RecoveryEmail}
 	user_infos = append(user_infos, user.LocalIdentities...)
@@ -37,10 +42,9 @@ func ChangeUserPassword(user *User, patch *gjson.Result, store backends.UserStor
 	// hash new password and store it
 	hashpass, err := bcrypt.GenerateFromPassword([]byte(new_password), defaultBcryptCost)
 	(*user).Password = hashpass
-	err = store.UpdateUserPassword(user)
+	err = store.UpdateUserPasswordHash(user)
 	if err != nil {
 		return errors.New("[ChangeUserPassword] failed to store updated user : " + err.Error())
 	}
-
 	return nil
 }